+++ src/pages/contents/mytree/items/MyTreeItemTree.jsx
@@ -1,0 +1,12 @@
+import React, { useEffect, useRef, useState } from 'react';
+import S from '../style';
+
+const MyTreeItemTree = () => {
+    return (
+        <div>
+            내 나무 돌리도
+        </div>
+    );
+};
+
+export default MyTreeItemTree;
